+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## 4.3.0 (2023-12-15)
+## 4.3.0 (2023-12-25)
 
 ### Features / Enhancements